Move immutable utils to utils/immutable submodule
