--- packages/immutable-store/README.md
@@ -1,38 +1,0 @@
-<p>
-  <img width="100%" src="https://assets.solidjs.com/banner?type=Primitives&background=tiles&project=immutable-store" alt="Solid Primitives immutable-store">
-</p>
-
-# @solid-primitives/immutable-store
-
-[![turborepo](https://img.shields.io/badge/built%20with-turborepo-cc00ff.svg?style=for-the-badge&logo=turborepo)](https://turborepo.org/)
-[![size](https://img.shields.io/bundlephobia/minzip/@solid-primitives/immutable-store?style=for-the-badge&label=size)](https://bundlephobia.com/package/@solid-primitives/immutable-store)
-[![version](https://img.shields.io/npm/v/@solid-primitives/immutable-store?style=for-the-badge)](https://www.npmjs.com/package/@solid-primitives/immutable-store)
-[![stage](https://img.shields.io/endpoint?style=for-the-badge&url=https%3A%2F%2Fraw.githubusercontent.com%2Fsolidjs-community%2Fsolid-primitives%2Fmain%2Fassets%2Fbadges%2Fstage-0.json)](https://github.com/solidjs-community/solid-primitives#contribution-process)
-
-A sample primitive that is made up for templating with the following options:
-
-`createPrimitiveTemplate` - Provides a getter and setter for the primitive.
-
-## Installation
-
-```bash
-npm install @solid-primitives/immutable-store
-# or
-yarn add @solid-primitives/immutable-store
-# or
-pnpm add @solid-primitives/immutable-store
-```
-
-## How to use it
-
-```ts
-const [value, setValue] = createPrimitiveTemplate(false);
-```
-
-## Demo
-
-You can use this template for publishing your demo on CodeSandbox: https://codesandbox.io/s/solid-primitives-demo-template-sz95h
-
-## Changelog
-
-See [CHANGELOG.md](./CHANGELOG.md)
