+++ packages/immutable/package.json
@@ -1,22 +1,32 @@
 {
   "name": "@solid-primitives/immutable",
-  "version": "0.1.10",
-  "description": "Functional programming helpers for making non-mutating changes to data. Keeping it immutable.",
-  "author": "Damian Tarnawski @thetarnav <gthetarnav@gmail.com>",
+  "version": "0.0.100",
+  "description": "A template primitive example.",
+  "author": "Damian Tarnawski <gthetarnav@gmail.com>",
+  "contributors": [],
   "license": "MIT",
   "homepage": "https://github.com/solidjs-community/solid-primitives/tree/main/packages/immutable#readme",
   "repository": {
     "type": "git",
     "url": "git+https://github.com/solidjs-community/solid-primitives.git"
   },
+  "bugs": {
+    "url": "https://github.com/solidjs-community/solid-primitives/issues"
+  },
   "primitive": {
     "name": "immutable",
-    "stage": 2,
+    "stage": 0,
     "list": [
-      "List of functions"
+      "createImmutable"
     ],
-    "category": "Utilities"
+    "category": "Reactivity"
   },
+  "keywords": [
+    "solid",
+    "primitives",
+    "immutable",
+    "store"
+  ],
   "private": false,
   "sideEffects": false,
   "files": [
@@ -34,22 +44,24 @@
     },
     "require": "./dist/index.cjs"
   },
+  "typesVersions": {},
   "scripts": {
+    "dev": "vite serve dev",
+    "page": "vite build dev",
     "build": "jiti ../../scripts/build.ts",
     "test": "vitest -c ../../configs/vitest.config.ts",
     "test:ssr": "pnpm run test --mode ssr"
   },
-  "keywords": [
-    "fp",
-    "immutable",
-    "solid",
-    "primitives"
-  ],
-  "dependencies": {
-    "@solid-primitives/utils": "workspace:^6.0.0"
-  },
   "peerDependencies": {
     "solid-js": "^1.6.12"
   },
-  "typesVersions": {}
+  "dependencies": {
+    "@solid-primitives/keyed": "workspace:^1.1.10",
+    "@solid-primitives/utils": "workspace:^6.0.0",
+    "redux": "^4.2.1"
+  },
+  "devDependencies": {
+    "@reduxjs/toolkit": "^1.9.3",
+    "solid-transition-group": "^0.2.2"
+  }
 }
