+++ packages/keyed/package.json
@@ -50,7 +50,6 @@
     "test:ssr": "pnpm run test --mode ssr"
   },
   "devDependencies": {
-    "@solid-primitives/immutable": "workspace:^0.1.10",
     "@solid-primitives/refs": "workspace:^1.0.2",
     "@solid-primitives/utils": "workspace:^6.0.0",
     "solid-transition-group": "^0.2.2"
