+++ packages/pointer/package.json
@@ -55,7 +55,6 @@
   },
   "dependencies": {
     "@solid-primitives/event-listener": "workspace:^2.2.10",
-    "@solid-primitives/immutable": "workspace:^0.1.10",
     "@solid-primitives/rootless": "workspace:^1.3.2",
     "@solid-primitives/utils": "workspace:^6.0.0"
   },
