+++ packages/refs/test/index.test.ts
@@ -1,6 +1,6 @@
 import { describe, test, expect } from "vitest";
 import { createRoot, createSignal } from "solid-js";
-import { removeItems } from "@solid-primitives/immutable";
+import { removeItems } from "@solid-primitives/utils/immutable";
 import { getResolvedElements, resolveElements } from "../src";
 
 const el1 = document.createElement("div");
