+++ packages/signal-builders/package.json
@@ -45,7 +45,6 @@
     "test:ssr": "pnpm run test --mode ssr"
   },
   "dependencies": {
-    "@solid-primitives/immutable": "workspace:^0.1.10",
     "@solid-primitives/utils": "workspace:^6.0.0"
   },
   "peerDependencies": {
