+++ packages/utils/src/immutable/types.ts
@@ -1,4 +1,4 @@
-import { ItemsOf } from "@solid-primitives/utils";
+import { ItemsOf } from "..";
 
 export type Predicate<T> = (item: T, index: number, array: readonly T[]) => boolean;
 export type MappingFn<T, V> = (item: T, index: number, array: readonly T[]) => V;
